test/cartpole/requirements.txt
atari-py==0.2.6
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
future==0.18.2
google-api-core==1.15.0
google-auth==1.9.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gym==0.15.4
idna==2.8
imageio==2.6.1
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.3.2
pytz==2019.3
requests==2.22.0
rsa==4.0
scipy==1.4.1
six==1.13.0
urllib3==1.25.7
keras
matplotlib
tensorflow